requirements.txt
pre-commit==3.3.3
black==23.3.0
isort==5.12.0
flake8==6.0.0
pandas==2.0.2
apache-airflow==2.6.2